- Added Index COLUMN_CURRENT_PRICE - Removed GMC task and keep it in the intent service to solve the problem of force close rebuild problem
