+++ app/src/main/AndroidManifest.xml
@@ -27,14 +27,6 @@
             android:authorities="br.com.carteira"
             android:exported="false" />
 
-        <service
-            android:name=".api.service.StockTaskService"
-            android:exported="true"
-            android:permission="com.google.android.gms.permission.BIND_NETWORK_TASK_SERVICE">
-            <intent-filter>
-                <action android:name="com.google.android.gms.gcm.ACTION_TASK_READY" />
-            </intent-filter>
-        </service>
         <service android:name=".api.service.StockIntentService" />
 
         <activity
